Move files, clean up imports and code
